Reticle Fixed
Reticle now turns red based on the range of the weapon while hovering over an enemy to give better player feedback

also restructured the crates hover methods to better match the rest of the hoverable interactions

+++ Assets/Scripts/_Core/Handlers/StatHandler.cs
@@ -32,7 +32,7 @@ public class StatHandler : MonoBehaviour, IDamage
 
 
     [Range(0f, 100f)][SerializeField] public float modSpeed;
-    [Range(0f, 100f)][SerializeField] public float modJumps;
+    [Range(0, 100)][SerializeField] public int modJumps;
 
     public Slider healthBar;
     public TMP_Text healthText;
